requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
bcrypt
passlib
python-dotenv
psycopg2-binary
